+++ src/pages/exercises/FashionMnist/components/index.ts
@@ -1,0 +1,3 @@
+export * from './InfoSection';
+export * from './TrainingSection';
+export * from './EvaluationSection';